boring_app/requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
requests==2.27.1
mysql.connector==2.2.9
